backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pymongo==4.6.0
pydantic==2.5.0
httpx==0.25.2
python-multipart==0.0.6
python-dateutil==2.8.2
telnyx==3.2.0
websockets==12.0
pyjwt==2.8.1
aiofiles==23.2.1
